Test against ASE calculator (currently failing?)
The difference between the two is fairly small but still significant.
I'm not sure where the discrepancy could be arising.
Visualising the component potentials, they look identical, so the fitting and parsing must be correct.
The issue then must lie in the actual energy evaluation. Could it be related to the neighbourlist?