Fix bug with species ordering
ASE provides the data in the order that matches the species list in the parameter file, and this is not necessarily in order.

Ordering the zlist then causes incorrect results. I've added the option to avoid sorting the zlist which fixes the problem. The test I've added was previously failing.